[TG-TECH][refactor]  refactored scenarios
