--- src/app/scenarios/index.js
@@ -1,8 +1,0 @@
-export * from './scenarios-list/scenario-list.component.js';
-export * from './scenarios-list/scenario-list.resolve.service';
-
-export * from './scenario-details/editor/edit-scenario-editor/edit-scenario-editor.component';
-export * from './scenario-details/editor/create-scenario-editor/create-scenario-editor.component';
-
-export * from './scenario-details/wizard/edit-scenario-wizard/edit-scenario-wizard.component';
-export * from './scenario-details/wizard/create-scenario-wizard/create-scenario-wizard.component';
