+++ src/app/scenarios/scenario-details/editor/create-scenario-editor/create-scenario-editor.template.html
@@ -24,7 +24,8 @@
     <label>
         Scenario name: <b class="red-text">*</b>
         <input [(ngModel)]="scenario.name"
-               #name="ngForm"
+               name="name"
+               #name="ngModel"
                class="black-text"
                type="text"
                required
@@ -36,6 +37,8 @@
 
     <p class="sensor-detail__status">
         <input [(ngModel)]="scenario.active"
+               name="active"
+               #active="ngModel"
                class="filled-in sensor-detail__status-value"
                type="checkbox" id="scenario-details__status"/>
         <label for="scenario-details__status">active </label>
@@ -44,6 +47,8 @@
     <label>
         Scenario description:
         <input [(ngModel)]="scenario.description"
+               name="description"
+               #description="ngModel"
                class="black-text"
                type="text"/>
     </label>
