+++ src/app/scenarios/scenario-details/editor/create-scenario-editor/scenario-editor.style.scss
