+++ src/app/scenarios/scenario-details/scenario-details.component.js
@@ -1,21 +1,18 @@
-import {Component} from '@angular/core';
-import { ROUTER_DIRECTIVES, ActivatedRoute, Router } from '@angular/router';
+import { Component } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 
-import template from './scenario-details.html';
-import style from './scenario-details.scss';
-import {ScenarioService} from '../shared/Scenario.service.js';
-import {ScenarioEditor} from './editor/scenario-editor/scenario-editor.component.js';
+import template from './scenario-details.template.html';
+import style from './scenario-details.style.scss';
+
+import { ScenarioService } from '../shared/scenario.service.js';
 
 const selector = 'scenario-list';
 
 @Component({
     selector,
     template,
-    styles: [style],
-    providers: [ScenarioService],
-    directives: [ROUTER_DIRECTIVES, ScenarioEditor]
+    styles: [style]
 })
-
 export class ScenarioDetailsComponent {
 
     constructor(scenarioService:ScenarioService, route:ActivatedRoute, router: Router) {
