+++ src/app/scenarios/scenario-details/scenario-details.style.scss
