+++ src/app/scenarios/scenario-details/scenario-details.template.html
