+++ src/app/scenarios/scenario-details/wizard/scenario-entities/action.js
