+++ src/app/scenarios/scenario-details/wizard/scenario-entities/condition.js
