+++ src/app/scenarios/scenario-details/wizard/scenario-entities/scenario.js
