--- src/app/scenarios/scenario-details/wizard/scenario-wizard/index.js
@@ -1,1 +1,0 @@
-export * from './scenario-wizard.component.js';
