+++ src/app/scenarios/scenario-details/wizard/scenario-wizard/scenario-wizard.style.scss
