+++ src/app/scenarios/scenarios-list/scenario-list.resolve.service.js
@@ -1,12 +1,11 @@
-import { ScenarioService } from '../shared/Scenario.service.js';
-import { Observable } from 'rxjs/Observable';
+import { ScenarioService } from '../shared/scenario.service.js';
 
 export class ScenarioListResolveService {
     constructor(scenarioService: ScenarioService) {
         this.scenarioService = scenarioService;
     }
 
-    resolve(): Observable<any> {
+    resolve() {
         return this.scenarioService.get();
     }
 }
