+++ src/app/scenarios/scenarios-list/scenario-list.style.scss
