+++ src/app/scenarios/scenarios-list/scenario-status.service.js
@@ -1,4 +1,4 @@
-import {Injectable} from '@angular/core';
+import { Injectable } from '@angular/core';
 
 import io from 'socket.io-client';
 import Rx from 'rxjs/Rx';
