+++ src/app/scenarios/shared/scenario.service.js
@@ -1,5 +1,5 @@
-import {Injectable} from '@angular/core';
-import ShHttpService from '../../shared/sh-http/sh-http.service.js';
+import { Injectable } from '@angular/core';
+import { ShHttpService } from '../../shared/sh-http/sh-http.service.js';
 
 // TODO: let's use 'streams' approach. All promises functions should be
 // removed in future after refactoring
